+++ resources/views/action_yaml.blade.php
@@ -15,7 +15,7 @@ jobs:
       matrix:
         operating-system: [ubuntu-latest]
         php-versions: {!! $stepPhpVersionsString !!}
-        dependency-stability: [ prefer-stable ]
+        dependency-stability: {!! $dependencyStabilityString !!}
 @if ($matrixLaravel)
 
         laravel: {!! $matrixLaravelVersionsString !!}
@@ -90,6 +90,15 @@ jobs:
     - name: Install Dependencies
       if: steps.vendor-cache.outputs.cache-hit != 'true'
       run: composer install -q --no-ansi --no-interaction --no-scripts --no-progress --prefer-dist
+@endif
+
+@if (is_array($dependencyStability) && count($dependencyStability) > 0)
+    - name: Update Dependencies with latest stable
+      if: matrix.dependency-stability == 'prefer-stable'
+      run: composer update --prefer-stable
+    - name: Update Dependencies with lowest stable
+      if: matrix.dependency-stability == 'prefer-lowest'
+      run: composer update --prefer-stable --prefer-lowest
 @endif
 
 @if ($stepGenerateKey)