+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.4.3 - WIP
 - Psalm, create Sarif report for GitHub code scanning integration
 - update to Alpinejs 3
+- Allow user to set dependency stability (none, the latest stable, the lowest stable)
 
 ## 0.4.2 - 2021-10-12
 Hacktoberfest, goes on...